feat(create-analog): add template for Angular v17 pre-release
