--- packages/create-analog/template-angular-v15/package.json
@@ -1,49 +1,0 @@
-{
-  "name": "my-app",
-  "version": "0.0.0",
-  "private": true,
-  "engines": {
-    "node": "^16.14.0 || >=18.10.0"
-  },
-  "scripts": {
-    "dev": "ng serve",
-    "ng": "ng",
-    "start": "ng serve",
-    "build": "ng build",
-    "watch": "ng build --watch",
-    "test": "ng test"
-  },
-  "dependencies": {
-    "@analogjs/content": "^0.2.0",
-    "@analogjs/router": "^0.2.0",
-    "@angular/animations": "^15.0.0",
-    "@angular/common": "^15.0.0",
-    "@angular/compiler": "^15.0.0",
-    "@angular/core": "^15.0.0",
-    "@angular/forms": "^15.0.0",
-    "@angular/platform-browser": "^15.0.0",
-    "@angular/platform-browser-dynamic": "^15.0.0",
-    "@angular/platform-server": "^15.0.0",
-    "@angular/router": "^15.0.0",
-    "@nrwl/angular": "^15.0.0",
-    "front-matter": "^4.0.2",
-    "marked": "^5.0.2",
-    "mermaid": "^10.2.4",
-    "prismjs": "^1.29.0",
-    "rxjs": "~7.5.6",
-    "tslib": "^2.4.0",
-    "zone.js": "~0.11.8"
-  },
-  "devDependencies": {
-    "@analogjs/platform": "0.1.0-beta.23",
-    "@angular-devkit/build-angular": "^15.0.0",
-    "@angular/cli": "~15.0.0",
-    "@angular/compiler-cli": "^15.0.0",
-    "@nrwl/vite": "^15.0.0",
-    "nx": "^15.7.0",
-    "jsdom": "^20.0.0",
-    "typescript": "~4.8.4",
-    "vite": "^4.0.3",
-    "vitest": "^0.25.8"
-  }
-}
