--- packages/create-analog/template-angular-v15/src/main.providers.ts
@@ -1,5 +1,0 @@
-/**
- * Common providers shared with client and server-side.
- */
-
-export const mainProviders = [];
