--- packages/create-analog/template-angular-v15/src/main.server.ts
@@ -1,26 +1,0 @@
-import 'zone.js/node';
-import { enableProdMode } from '@angular/core';
-import { renderApplication } from '@angular/platform-server';
-import { provideFileRouter } from '@analogjs/router';
-import { withEnabledBlockingInitialNavigation } from '@angular/router';
-
-import { AppComponent } from './app/app.component';
-import { mainProviders } from './main.providers';
-
-if (import.meta.env.PROD) {
-  enableProdMode();
-}
-
-export default async function render(url: string, document: string) {
-  const html = await renderApplication(AppComponent, {
-    appId: 'analog-app',
-    document,
-    url,
-    providers: [
-      provideFileRouter(withEnabledBlockingInitialNavigation()),
-      ...mainProviders,
-    ],
-  });
-
-  return html;
-}
