+++ packages/create-analog/template-angular-v17-prerelease/.vscode/launch.json
