+++ packages/create-analog/template-angular-v17-prerelease/_gitignore
@@ -30,6 +30,7 @@ yarn-error.log
 
 # Miscellaneous
 /.angular/cache
+/.nx/cache
 .sass-cache/
 /connect.lock
 /coverage
