+++ packages/create-analog/template-angular-v17-prerelease/angular.json
@@ -10,11 +10,12 @@
       "prefix": "app",
       "architect": {
         "build": {
-          "builder": "@nrwl/vite:build",
+          "builder": "@analogjs/platform:vite",
           "options": {
             "configFile": "vite.config.ts",
             "main": "src/main.ts",
-            "outputPath": "dist/client"
+            "outputPath": "dist/client",
+            "tsConfig": "tsconfig.app.json"
           },
           "defaultConfiguration": "production",
           "configurations": {
@@ -28,7 +29,7 @@
           }
         },
         "serve": {
-          "builder": "@nrwl/vite:dev-server",
+          "builder": "@analogjs/platform:vite-dev-server",
           "defaultConfiguration": "development",
           "options": {
             "buildTarget": "my-app:build",
@@ -45,7 +46,7 @@
           }
         },
         "test": {
-          "builder": "@nrwl/vite:test"
+          "builder": "@analogjs/platform:vitest"
         }
       }
     }
