+++ packages/create-analog/template-angular-v17-prerelease/favicon.ico
