+++ packages/create-analog/template-angular-v17-prerelease/package.json
@@ -1,0 +1,51 @@
+{
+  "name": "my-app",
+  "version": "0.0.0",
+  "private": true,
+  "engines": {
+    "node": ">=18.13.0"
+  },
+  "scripts": {
+    "dev": "ng serve",
+    "ng": "ng",
+    "start": "ng serve",
+    "build": "ng build",
+    "watch": "ng build --watch",
+    "test": "ng test"
+  },
+  "dependencies": {
+    "@analogjs/content": "^0.2.0",
+    "@analogjs/router": "^0.2.0",
+    "@angular/animations": "^17.0.0-rc.0",
+    "@angular/common": "^17.0.0-rc.0",
+    "@angular/compiler": "^17.0.0-rc.0",
+    "@angular/core": "^17.0.0-rc.0",
+    "@angular/forms": "^17.0.0-rc.0",
+    "@angular/platform-browser": "^17.0.0-rc.0",
+    "@angular/platform-browser-dynamic": "^17.0.0-rc.0",
+    "@angular/platform-server": "^17.0.0-rc.0",
+    "@angular/router": "^17.0.0-rc.0",
+    "@nx/angular": "^17.0.0",
+    "front-matter": "^4.0.2",
+    "marked": "^5.0.2",
+    "marked-gfm-heading-id": "^3.1.0",
+    "marked-highlight": "^2.0.1",
+    "mermaid": "^10.2.4",
+    "prismjs": "^1.29.0",
+    "rxjs": "~7.5.6",
+    "tslib": "^2.4.0",
+    "zone.js": "~0.14.0"
+  },
+  "devDependencies": {
+    "@analogjs/platform": "^0.2.0",
+    "@angular-devkit/build-angular": "^17.0.0-rc.0",
+    "@angular/cli": "^17.0.0-rc.0",
+    "@angular/compiler-cli": "^17.0.0-rc.0",
+    "@nx/vite": "~16.8.1",
+    "nx": "^17.0.0",
+    "jsdom": "^22.1.0",
+    "typescript": "~5.2.0",
+    "vite": "^4.4.8",
+    "vitest": "^0.32.0"
+  }
+}
