+++ packages/create-analog/template-angular-v17-prerelease/src/app/app.component.ts
