+++ packages/create-analog/template-angular-v17-prerelease/src/app/app.config.server.ts
@@ -1,0 +1,16 @@
+import { mergeApplicationConfig, ApplicationConfig } from '@angular/core';
+import {
+  provideServerRendering,
+  ɵSERVER_CONTEXT as SERVER_CONTEXT,
+} from '@angular/platform-server';
+
+import { appConfig } from './app.config';
+
+const serverConfig: ApplicationConfig = {
+  providers: [
+    provideServerRendering(),
+    { provide: SERVER_CONTEXT, useValue: 'ssr-analog' },
+  ],
+};
+
+export const config = mergeApplicationConfig(appConfig, serverConfig);
