+++ packages/create-analog/template-angular-v17-prerelease/src/app/app.config.ts
@@ -1,0 +1,12 @@
+import { provideHttpClient, withFetch } from '@angular/common/http';
+import { ApplicationConfig } from '@angular/core';
+import { provideClientHydration } from '@angular/platform-browser';
+import { provideFileRouter } from '@analogjs/router';
+
+export const appConfig: ApplicationConfig = {
+  providers: [
+    provideFileRouter(),
+    provideHttpClient(withFetch()),
+    provideClientHydration(),
+  ],
+};
