+++ packages/create-analog/template-angular-v17-prerelease/src/app/pages/index.page.ts
