+++ packages/create-analog/template-angular-v17-prerelease/src/main.ts
@@ -1,0 +1,7 @@
+import 'zone.js';
+import { bootstrapApplication } from '@angular/platform-browser';
+
+import { AppComponent } from './app/app.component';
+import { appConfig } from './app/app.config';
+
+bootstrapApplication(AppComponent, appConfig);
