+++ packages/create-analog/template-angular-v17-prerelease/src/server/routes/v1/hello.ts
