+++ packages/create-analog/template-angular-v17-prerelease/src/vite-env.d.ts
