+++ packages/create-analog/template-angular-v17-prerelease/tsconfig.app.json
