+++ packages/create-analog/template-angular-v17-prerelease/vite.config.ts
